Move out AboutDialog out of ui.py
